End point for returning places of interest
We have a KML of public art pieces around MK so we will serve
up the contents of this to show on the map.

The PlacesController has a list of PlaceOfInterestProviders
injected into it. If you define a bean of this type then
it will automatically be injected in and the places will
start being served up.